+++ bundled/minimal/src/minimal_clj/user.clj
@@ -1,1 +1,0 @@
-(ns minimal-clj.user)