+++ src/nrepl/project-types.ts
@@ -124,16 +124,20 @@ async function selectShadowBuilds(
   connectSequence: ReplConnectSequence,
   foundBuilds: string[]
 ): Promise<{ selectedBuilds: string[]; args: string[] }> {
-  const menuSelections = connectSequence.menuSelections,
-    selectedBuilds = menuSelections
-      ? menuSelections.cljsLaunchBuilds
-      : await utilities.quickPickMulti({
-          values: foundBuilds.filter((x) => x[0] == ':').map((a) => ({ label: a })),
-          placeHolder: 'Select builds to start',
-          saveAs: `${state.getProjectRootUri().toString()}/shadow-cljs-jack-in`,
-        }),
-    aliases: string[] =
-      menuSelections && menuSelections.cljAliases ? menuSelections.cljAliases.map(keywordize) : [];
+  const menuSelections = connectSequence.menuSelections;
+  let selectedBuilds: string[];
+  if (menuSelections && menuSelections.cljsLaunchBuilds) {
+    selectedBuilds = menuSelections.cljsLaunchBuilds;
+  } else {
+    const selectedBuildItems = await utilities.quickPickMulti({
+      values: foundBuilds.filter((x) => x[0] == ':').map((a) => ({ label: a })),
+      placeHolder: 'Select builds to start',
+      saveAs: `${state.getProjectRootUri().toString()}/shadow-cljs-jack-in`,
+    });
+    selectedBuilds = selectedBuildItems.map((item) => item.label);
+  }
+  const aliases: string[] =
+    menuSelections && menuSelections.cljAliases ? menuSelections.cljAliases.map(keywordize) : [];
   const aliasesOption = aliases.length > 0 ? ['-A', aliases.join('')] : [];
   const args: string[] = [];
   if (aliasesOption && aliasesOption.length) {
@@ -182,11 +186,12 @@ async function leinProfilesAndAlias(
           aliases = Object.keys(aliasesMap).map((v, k) => v);
           if (aliases.length) {
             aliases.unshift('No alias');
-            alias = await utilities.quickPickSingle({
+            const aliasItems = await utilities.quickPickSingle({
               values: aliases.map((a) => ({ label: a })),
               saveAs: `${state.getProjectRootUri().toString()}/lein-cli-alias`,
               placeHolder: 'Choose alias to launch with',
             });
+            alias = aliasItems?.label;
             alias = alias == 'No alias' ? undefined : alias;
           }
         }
@@ -210,11 +215,12 @@ async function leinProfilesAndAlias(
       if (projectProfiles.length) {
         profiles = projectProfiles.map(keywordize);
         if (profiles.length) {
-          profiles = await utilities.quickPickMulti({
+          const profilItems = await utilities.quickPickMulti({
             values: profiles.map((a) => ({ label: a })),
             saveAs: `${state.getProjectRootUri().toString()}/lein-cli-profiles`,
             placeHolder: 'Pick any profiles to launch with',
           });
+          profiles = profilItems.map((item) => item.label);
         }
       }
     }
@@ -628,7 +634,7 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
       projectAliases = [...projectAliases, ...myAliases];
     }
     if (projectAliases.length) {
-      aliases = await utilities.quickPickMulti({
+      const aliasItems = await utilities.quickPickMulti({
         values: projectAliases
           .map(keywordize)
           .sort()
@@ -644,6 +650,7 @@ async function cljCommandLine(connectSequence: ReplConnectSequence, cljsType: Cl
         saveAs: `${state.getProjectRootUri().toString()}/clj-cli-aliases`,
         placeHolder: 'Pick any aliases to launch with',
       });
+      aliases = aliasItems?.map((item) => item.label);
     }
   }
   const selectedAliasesHasMain = aliases.filter((a) => aliasesWithMain.includes(a)).length > 0;